==> data+analysis/07_main_contextSensitive_pilot_analysis.Rmd <==
---
title: "QA free typing with varying context pilot analysis"
author: "Polina Tsvilodub"
date: '2023-01-14'
output: github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r, include=FALSE}
library(tidyverse)
library(tidyboot)
library(aida)
library(brms)
library(tidybayes)
library(cspplot)
```

```{r, include=FALSE}
# these options help Stan run faster
options(mc.cores = parallel::detectCores())
```

## Intro

The vignettes presented pairs of different contexts. However, the list of alternatives was the same for both contexts. The hypothesis is that the suggested competitors will differ across contexts.

The responses were manually classified by Polina into the following categories: 

* competitor: responses mentioning the anticipated competitor only (vary across contexts, sometimes several options were categorized as competitors)
* mostSimilar: one of the sameCategory alternatives is an items that can be considered most similar to the target _out of context_ (i.e., same for both contexts), so this category comprises responses offering that alternative
* sameCategory: responses offering both same category alternatives or offering the option which we did not consider the direct competitor (options categorised as competitors in one context usually are sameCategory items in the second context)
* otherCategory: responses offering the alternative from the different category (distractors; usually same in both contexts. In some vignettes, there are two, in some there is only one other category alternative)
* fullList: responses where all alternatives were listed (also in two sentences, where one offered the competitor only etc)
* taciturn: responses not offering any alternative options or further alternative solutions
* other: responses where a same category + other category response are mixed, uncertain answers, unclassifiable responses, responses offering further teps towards solcing the problem, responses using basic level categories (e.g., "dogs" instead of offering specific alternatives)
* yes responses (excluded from analyses when not mentioning one of the alternatives)


In the experiment, each subject saw *four main trials* and *one attention checking trial*. Participants failing attention checks were excluded from analysis. They were reimbursed 0.60 pounds for their participation.

## Summary and exclusions

```{r, include=FALSE, warning=FALSE, message=FALSE}
answerOrder <- c( 'competitor', 'mostSimilar', 'sameCategory', 'otherCategory', 'fullList', 'taciturn', 'other')
answerOrder_rev <- c('other', 'taciturn', 'fullList',  'otherCategory', 'sameCategory', 'mostSimilar','competitor')
#df <- read_csv("../../raw_data/results_103_QA-overinfo-freeTyping-cogsci_PT_full.csv")
#df %>% select(-prolific_pid, -prolific_session_id, -prolific_study_id) %>%  write_csv("data/results_QA-overinfo-freeTyping-cogsci_full_anonymized.csv")

df <- read_csv("data/results_105_QA-overinfo-contextDependent-freeTyping-pilot4_130_categorized.csv") %>% filter((itemName != "chair-repair") & (itemName != "chair-party"))
#helper <- read_csv("../../raw_data/results_105_QA-overinfo-contextDependent-freeTyping-pilot4_130.csv")

#joined_df<- helper %>% left_join(., df %>% select(submission_id, answer, itemName, response_option, category), by = c("submission_id", "answer", "itemName") ) %>% select(-prolific_pid, prolific_session_id, -prolific_study_id)
#joined_df %>% write_csv("data/results_105_QA-overinfo-contextDependent-freeTyping-pilot4_130_categorized.csv")

head(df)
cat("Number of recruited subjects: ", df %>% pull(submission_id) %>% unique() %>% length())
```

```{r, echo=FALSE}
# check attention check completion
df_attention <- df %>% filter(trial_type == "filler") %>% rowwise() %>%
  mutate(passed_trial = grepl(tolower(correct_response), tolower(answer))) 

df_attention_fail <- df_attention %>% group_by(submission_id) %>% 
  mutate(passed_subj = sum(passed_trial) > 0)

# participants who failed attention checks 
subj_id_attention_fails <- df_attention_fail %>% filter(passed_subj == FALSE) %>% pull(submission_id) %>% unique()
cat("Subjects who failed attention checks: ", subj_id_attention_fails)  
cat("\nNumber of subjects who failed attention checks: ", length(subj_id_attention_fails))
cat("\nSubject exclusion rate: ", length(subj_id_attention_fails)/(unique(df$submission_id) %>% length()))
```

```{r, echo=FALSE}
cat("Number of excluded inccorect ('yes') responses: ", nrow(df %>% filter(category == "yes")))
cat("Proportion of excluded inccorect ('yes') responses: ", nrow(df %>% filter(category == "yes"))/nrow(df))

# exclude failing participants and incorrect ("yes") responses
df_clean_main <- df %>% filter(!(submission_id %in% subj_id_attention_fails)) %>%
  filter(trial_type == "main", category != "yes") %>%
  mutate(category = factor(category))

#df_clean_main %>% select(submission_id, answer, itemName, settingName, category, response_option) %>% write_csv("../code/paper_data/e2_human_category_raw.csv")
cat("\nNumber of subjects post exclusions: ", df_clean_main %>% pull(submission_id) %>% unique() %>% length())
cat("\nNumber of analysed responses: ", nrow(df_clean_main))

df_clean_main %>% count(category)

df_clean_main %>% count(itemName) 

cat("\naverage number of responses per vignette:", mean(df_clean_main %>% count(itemName) %>% pull(n)))

cat("\nvignette with most responses: ", df_clean_main %>% count(itemName) %>% arrange(desc(n)) %>% .[1,] %>% .$itemName, df_clean_main %>% count(itemName) %>% arrange(desc(n)) %>% .[1,] %>% .$n)
cat("\nvignette with least responses: ", df_clean_main %>% count(itemName) %>% arrange(n) %>% .[1,] %>% .$itemName, df_clean_main %>% count(itemName) %>% arrange(n) %>% .[1,] %>% .$n)
```
```{r, echo=FALSE, message=FALSE, warning=FALSE}
e3_pilot_vignettes <- read_csv("../experiments/contextSensitive_free_production/trials/trials_split_cogsci_pilot4.csv") %>% 
  select(itemName, settingName, itemQuestion, competitor, mostSimilar, sameCategory,otherCategory) %>%
  mutate(context_nr = rep(c("context1", "context2"), 13))
df_clean_main %>% count(itemName) %>% pull(n) %>% mean()
e3_pilot_wVignette <- left_join(df_clean_main, e3_pilot_vignettes, by = c("itemName", "settingName")) 

e3_summary_byItem <- df_clean_main %>% group_by(itemName) %>%
  mutate(response_count = n())

e3_summary_byItem <- e3_summary_byItem %>% group_by(itemName, category) %>% 
  mutate(responseCategory_count = n(),
         responseCategory_proportion = responseCategory_count / response_count
         ) %>% ungroup() %>%
  select(itemName, category, responseCategory_proportion) %>% unique() 

e3_summary <- df_clean_main %>% ungroup() %>% group_by(category) %>% 
  mutate(responseCategory_count = n(),
         responseCategory_proportion = responseCategory_count / nrow(df_clean_main)
         ) %>% ungroup() %>%
  select(category, responseCategory_proportion) %>% unique() 

e3_pilot_items_cats <- read_csv("../experiments/contextSensitive_free_production/trials/trials_split_cogsci_pilot4_ItemCategorization.csv") %>%
  select(-itemName, -itemQuestion)  

e3_pilot_wVignette_raw <- e3_pilot_wVignette %>% filter((response_option != "yes") & (response_option != "no") & (response_option != "other")) %>%
  group_by(context_nr) %>%
  mutate(context_count = n()) %>%
  separate_rows(response_option, sep=",") %>%
  mutate(response_option = gsub(" ", "", response_option)) %>%
  #pivot_wider(values_from = category, names_from = context_nr) 
  left_join(., e3_pilot_items_cats, by = c("settingName", "response_option"))

e3_summary_subjMentions <- e3_pilot_wVignette_raw %>% ungroup() %>% group_by(global_category, itemName) %>%
  mutate(responseCategory_count = n()) %>% ungroup() %>% 
  mutate(subjMention_prop = responseCategory_count / vignette_count)

e3_pilot_wVignette_raw_subjMention_wide <- e3_summary_subjMentions %>%
  group_by(context_nr, global_category) %>%
  mutate(response_prop = mean(subjMention_prop)
         ) %>% select(settingName, context_nr, category, global_category, response_prop) %>%
  unique() %>%
  pivot_wider(values_from = response_prop, names_from = context_nr) %>%
  mutate(context1 = ifelse(is.na(context1), 0, context1),
         context2 = ifelse(is.na(context2), 0, context2)
         ) %>% group_by(global_category) %>%
  mutate(mean_context1 = mean(context1),
         mean_context2 = mean(context2)) %>% select(-settingName, -context1, -context2, -category) %>% unique() %>%
  rename("context1" = "mean_context1", "context2"="mean_context2")

e3_pilot_wVignette_raw_subjMention_wide

e3_pilot_wVignette_raw_subjMention_wide_acrossCats <- e3_summary_subjMentions %>% 
  select(settingName, context_nr, category, global_category, subjMention_prop) %>%
  unique() %>%
  group_by(global_category, context_nr) %>%
  mutate(mean_subjMention_prop = mean(subjMention_prop)) %>%
  select(context_nr, global_category, mean_subjMention_prop) %>% unique() %>%
  pivot_wider(values_from = mean_subjMention_prop, names_from = context_nr) 
 
```



```{r}
e3_summary_subjMentions %>% 
  select(context_nr, global_category, vignette_count, responseCategory_count) %>% 
  unique() %>% 
  arrange(context_nr, global_category) %>% 
  group_by(context_nr, global_category) %>% 
  summarize(X = sum(responseCategory_count) / sum(vignette_count))



e3_pilot_wVignette_raw %>% select(itemName, context_nr, global_category, submission_id, trialNr, context_count) %>% 
  arrange(submission_id, trialNr) %>% 
  group_by(context_nr, global_category) %>% summarize(x = n() / context_count) %>% unique() %>%
  pivot_wider(values_from = "x", names_from = "context_nr") %>% mutate(prompt = "human E2") %>%
  write_csv("../code/paper_data/e2_human_props_fixed.csv")

```


```{r}
e3_summary_subjMentions_context1 <- e3_summary_subjMentions %>% filter(context_nr == "context1") %>%
  group_by(global_category) %>%
  summarize(count = n(),
            prop = count / nrow(e3_summary_subjMentions %>% filter(context_nr == "context1"))
            )

e3_summary_subjMentions_context2 <- e3_summary_subjMentions %>% filter(context_nr == "context2") %>%
  group_by(global_category) %>%
  summarize(count = n(),
            prop = count / nrow(e3_summary_subjMentions %>% filter(context_nr == "context2"))
            )
```


Below, the response type patterns are reported. The types are classified akin to Experiment 1, with the addition of the mostSimilar category.
First, the response category distribution across vignettes is reported.
```{r}
e3_summary %>% 
  mutate(category = factor(category, levels = answerOrder_rev,
                           labels = c("other", "no option", "all options", "unrelated option\n(broom)","similar option\n(stool)", "related option\n(recliner)", "competitor\n(ladder)")
                           )) %>%
  ggplot(., aes(x = category, y = responseCategory_proportion, fill = category)) +
  geom_col(color = "#575463", fill = "#FFFFFF") +
  theme_csp() +
  theme(legend.position = "none", plot.title = element_text(hjust = 0.5)) +
  #theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  ylab("Proportion of answer type") +
  xlab("") +
  labs(fill = "Answer type") + coord_flip() +
  ggtitle("Do you have a chair?")
  

#ggsave("viz/e2_free_production_global_answerTypes_cogsci_vertical.pdf", width = 5, height = 4)
```

```{r,  fig.height=25, fig.width=4, echo=FALSE, eval=FALSE}
#  Pilot 4 with all items, plot by-vignette.
e3_pilot_wVignette_filtered <- e3_pilot_wVignette_raw %>% count(settingName, context_nr, itemName, response_option) %>%
  pivot_wider(values_from = n, names_from = context_nr) %>%
  mutate(context1 = ifelse(is.na(context1), 0, context1),
         context2 = ifelse(is.na(context2), 0, context2),
         ) 
e3_pilot_wVignette_filtered %>%
  ggplot(., aes(x = response_option)) +
  geom_col(aes(y = context1)) +
  geom_col(aes(y=-context2)) +
  facet_wrap(settingName ~., ncol= 1, scales = "free_y") +
  theme_csp() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  geom_hline(yintercept = 0) +
  theme(panel.spacing = unit(3, "lines")) +
  coord_flip() +
  ylab("")
```

Below, context 1 vs context 2 are compared in terms of numbers of mentions of the different option types in the respective context (comp_c1, comp_c2, mostSimilar, otherCategory; mostSimilar and otherCategory match for both contexts ). The mapping of alternatives to these option types depending on context can be found in the experiment folder file. To this end, the responses which do not mention any alternatives are excluded (i.e., other category, yes, and no are excluded).

Below, a plot of *absolute differences* in production proportions of the different response options are shown. This is so as to not to cancel out the plotted proportions with what is categorized as context 1 vs context 2 when comparing the categorization.
```{r, echo=FALSE}
e3_pilot_wVignette_raw_summary <- e3_pilot_wVignette_raw %>% group_by(settingName, context_nr, global_category) %>%
  mutate(responseCat_count = n()) %>% ungroup() %>% group_by(settingName, context_nr) %>%
  mutate(response_count = n()) %>% ungroup() %>%
  mutate(response_prop = responseCat_count / response_count
         ) %>% select(settingName, context_nr, global_category, response_prop) %>%
  unique() %>%
  pivot_wider(values_from = response_prop, names_from = context_nr) %>%
  mutate(context1 = ifelse(is.na(context1), 0, context1),
         context2 = ifelse(is.na(context2), 0, context2),
         abs_diff = abs(context1 - context2)
         ) %>% group_by(global_category) %>%
  summarize(mean_abs_diff = mean(abs_diff))
```

```{r, echo=FALSE}
e3_pilot_wVignette_raw_summary %>%
  ggplot(., aes(fill = global_category, x = global_category)) +
  geom_col(aes(y = mean_abs_diff)) +
  theme_csp() +
  #geom_col(aes(y = -context2)) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  #geom_hline(yintercept =0) +
  #coord_flip() +
  ylab("Absolute difference in\n mentioning proportions") +
  xlab("Alternative category")
```

Below, the distribution comparison of response proportions mirrored by context is displayed. 

```{r, echo=FALSE, fig.width = 6, fig.height = 3}
e3_pilot_wVignette_raw_byContext <- e3_pilot_wVignette_raw %>% group_by(settingName, context_nr, global_category) %>%
  mutate(responseCat_count = n()) %>% ungroup() %>% group_by(settingName, context_nr) %>%
  mutate(response_count = n()) %>% ungroup() %>%
  mutate(response_prop = responseCat_count / response_count
         ) %>% select(settingName, context_nr, category, global_category, response_prop) %>%
  unique() %>%
  pivot_wider(values_from = response_prop, names_from = context_nr) %>%
  mutate(context1 = ifelse(is.na(context1), 0, context1),
         context2 = ifelse(is.na(context2), 0, context2)
         ) %>% group_by(global_category) %>%
  mutate(mean_context1 = mean(context1),
         mean_context2 = mean(context2)) %>% select(-settingName, -context1, -context2, -category) %>% unique()

e3_pilot_wVignette_raw_byContext %>%
  ggplot(., aes(fill = global_category, x = global_category)) +
  geom_col(aes(y = mean_context1)) +
  geom_col(aes(y = -mean_context2)) +
  theme_csp() +
 theme(axis.text.y = element_blank(), axis.ticks.y = element_blank(), legend.position = "top") +
  scale_y_continuous( limits = c(-0.6, 0.6)) +
  geom_hline(yintercept =0, color = "white", size = 2) +
  coord_flip() +
  ylab("context 2 <--> context 1") +
  xlab("Alternative category")
```

Look at alternative mentioning proportions by-item:
```{r, echo=FALSE, fig.height=22, fig.width=7}
e3_pilot_wVignette_raw %>%
  group_by(settingName, itemName, response_option) %>%
  mutate(responseCat_count = n()) %>% ungroup() %>% group_by(settingName, context_nr) %>%
  mutate(response_count = n()) %>% ungroup() %>%
  mutate(response_prop = responseCat_count / response_count
         ) %>% select(settingName, context_nr, global_category, response_option, response_prop) %>% unique() %>%
#  mutate(n = n()) %>%
  pivot_wider(values_from = response_prop, names_from = context_nr) %>%
  mutate(context1 = ifelse(is.na(context1), 0, context1),
         context2 = ifelse(is.na(context2), 0, context2),
         ) %>%
  ggplot(., aes(x = response_option, fill = global_category)) +
  geom_col(aes(y = context1)) +
  geom_col(aes(y=-context2)) +
  facet_wrap(settingName ~., ncol= 2, scales = "free_y") +
  scale_y_continuous( limits = c(-1, 1)) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  geom_hline(yintercept = 0) +
  theme(panel.spacing = unit(3, "lines")) +
  coord_flip() +
  ylab("context 2 <--> context 1")
```

```{r, echo=FALSE, message=FALSE, warning=FALSE}
# plot suggested by Robert
e3_pilot_wVignette_raw_subjMention_wide <- e3_summary_subjMentions %>%
  group_by(context_nr, global_category) %>%
  mutate(response_prop = mean(subjMention_prop)
         ) %>% select(settingName, context_nr, category, global_category, response_prop) %>%
  unique() %>%
  pivot_wider(values_from = response_prop, names_from = context_nr) %>%
  mutate(context1 = ifelse(is.na(context1), 0, context1),
         context2 = ifelse(is.na(context2), 0, context2)
         ) %>% group_by(global_category) %>%
  mutate(mean_context1 = mean(context1),
         mean_context2 = mean(context2)) %>% select(-settingName, -context1, -context2, -category) %>% unique() %>%
  rename("context1" = "mean_context1", "context2"="mean_context2")

#e3_pilot_wVignette_raw_subjMention_wide %>% write_csv("../code/paper_data/e2_human_summary_wide.csv")

e3_pilot_wVignette_raw_subjMention_wide %>%
  mutate(global_category = factor(global_category, levels = c("otherCategory", "mostSimilar", "competitor_c2", "competitor_c1"), 
                                  labels = c("unrelated option (broom)", "similar option (recliner)", "context 2 competitor (stool)", "context 1 competitor (ladder)"))) %>%
  ggplot(., aes(fill = global_category, x = global_category)) +
  geom_col(aes(y = mean_context1)) +
  geom_col(aes(y = -mean_context2)) +
  theme_csp() + 
  scale_fill_manual(values = c("#E2BA78", "#99C2C2", "#C65353", "#7581B3")) +
  theme(legend.position = "none", axis.text.y = element_text(size = 14), axis.text.x = element_text(size = 14)) + # , axis.ticks.y = element_blank(), 
  scale_y_continuous( limits = c(-1, 1), breaks = c(-1, -0.5, 0, 0.5, 1), labels = c("1", "0.5", "0", "0.5", "1")) +
  geom_hline(yintercept = 0, color = "white", size = 1) +
  coord_flip() +
  ylab("context 2 <--> context 1") +
  xlab("Alternative category")

#ggsave("viz/e2_byContext_mentionProps.pdf", width = 6, height = 4)
```

## Exploratory stats

```{r}
multinomial_context_brm <- brm(
  global_category ~ 1 + context_nr + (1|settingName),
  data =  e3_pilot_wVignette_raw,
  family = "categorical",
  iter = 3000,
  control = list(adapt_delta=0.96),
  cores = 4
)
summary(multinomial_context_brm)
```
```{r}
multinom_context_posteriors <- multinomial_context_brm %>% spread_draws(b_mucompetitorc2_Intercept, b_mumostSimilar_Intercept, b_muotherCategory_Intercept, b_mucompetitorc2_context_nrcontext2, b_mumostSimilar_context_nrcontext2, b_muotherCategory_context_nrcontext2) %>%
  mutate(
    comp_c2 = b_mucompetitorc2_context_nrcontext2 - b_mucompetitorc2_Intercept,
    most_similar = b_mumostSimilar_context_nrcontext2 - b_mumostSimilar_Intercept,
    other_cat =  b_muotherCategory_context_nrcontext2 - b_muotherCategory_Intercept,
    comp_c1 = (exp(b_mucompetitorc2_context_nrcontext2) + exp(b_mumostSimilar_context_nrcontext2) + exp(b_muotherCategory_context_nrcontext2)) - (exp(b_mucompetitorc2_Intercept) + exp(b_mumostSimilar_Intercept) + exp(b_muotherCategory_Intercept))
  )

multinom_context_posteriors %>% select(comp_c2, most_similar, other_cat, comp_c1) %>%
  gather(key, val) %>%
  group_by(key) %>%
  summarize(
    '|95%' = quantile(val, probs = c(0.025, 0.975))[[1]],
    'mean'  = mean(val),
    '95%|' = quantile(val, probs = c(0.025, 0.975))[[2]],
    prob_gt_0 = mean(val > 0)*100,
    prob_lt_0 = mean(val < 0)*100
  ) -> multinom_context_posteriors_summary

```

```{r}
multinomial_e2_brm <- brm(
  category ~ 1,
  data =  df_clean_main,
  family = "categorical",
  iter = 3000,
  control = list(adapt_delta=0.9),
  cores = 4
)
summary(multinomial_e2_brm)#$fixeff
#fixef(multinomial_e2_brm)[,1]

```

```{r}
multinom_e2_posteriors <- multinomial_e2_brm %>% spread_draws(b_musameCategory_Intercept, b_muotherCategory_Intercept, b_mufullList_Intercept, b_mumostSimilar_Intercept, b_mutaciturn_Intercept) %>%
  mutate(
    sameCategory_vs_fullList = b_musameCategory_Intercept - b_mufullList_Intercept,
    sameCategory_vs_otherCategory = b_musameCategory_Intercept - b_muotherCategory_Intercept,
    sameCategory_vs_mostSimilar = b_musameCategory_Intercept - b_mumostSimilar_Intercept,
    sameCategory_vs_taciturn = b_musameCategory_Intercept - b_mutaciturn_Intercept
  )

multinom_e2_posteriors %>% select(sameCategory_vs_fullList, sameCategory_vs_otherCategory, sameCategory_vs_mostSimilar, sameCategory_vs_taciturn) %>%
  gather(key, val) %>%
  group_by(key) %>%
  summarize(
    '|95%' = quantile(val, probs = c(0.025, 0.975))[[1]],
    'mean'  = mean(val),
    '95%|' = quantile(val, probs = c(0.025, 0.975))[[2]],
    prob_gt_0 = mean(val > 0)*100,
    prob_lt_0 = mean(val < 0)*100
  ) -> multinom_e2_posteriors_summary

multinom_e2_posteriors_summary
```
